+++ NEWS.md
@@ -10,6 +10,8 @@
 * Row selection columns now display correctly in tables with column groups
   ([#52](https://github.com/glin/reactable/issues/52)).
 * `defaultSelected` now works correctly with Crosstalk linked selection.
+* Crosstalk selection and filtering now works with nested and dynamically
+  rendered tables [#57](https://github.com/glin/reactable/issues/57)).
 * Tables now display correctly for Crosstalk `SharedData` objects with zero or one rows.
 * Shiny UI elements in expanded row details are now properly removed when
   collapsed on page changes.